+++ packages/atomic/cypress/integration/facets/facet/facet-selectors.ts
@@ -18,46 +18,58 @@ export const FacetSelectors = {
     return this.shadow().find('[part="placeholder"]');
   },
   selectedCheckboxValue() {
-    return this.shadow().find('[part="value-checkbox"][aria-checked="true"]');
+    return this.shadow().find(
+      '[part~="value-checkbox"][part~="value-checkbox-checked"][aria-checked="true"]'
+    );
   },
   idleCheckboxValue() {
-    return this.shadow().find('[part="value-checkbox"][aria-checked="false"]');
+    return this.shadow().find(
+      '[part~="value-checkbox"]:not([part~="value-checkbox-checked"])[aria-checked="false"]'
+    );
   },
   checkboxValueWithText(text: string) {
     return this.shadow()
       .find(`[part="value-label"]:contains("${text}")`)
       .parent()
       .parent()
-      .find('[part="value-checkbox"]');
+      .find('[part~="value-checkbox"]');
   },
   idleCheckboxValueLabel() {
     return this.idleCheckboxValue().parent().find('[part="value-label"]');
   },
   selectedLinkValue() {
-    return this.shadow().find('[part="value-link"][aria-pressed="true"]');
+    return this.shadow().find(
+      '[part~="value-link"][part~="value-link-selected"][aria-pressed="true"]'
+    );
   },
   idleLinkValue() {
-    return this.shadow().find('[part="value-link"][aria-pressed="false"]');
+    return this.shadow().find(
+      '[part~="value-link"]:not([part~="value-link-selected"])[aria-pressed="false"]'
+    );
   },
   selectedLinkValueWithText(text: string) {
     return this.shadow().find(
-      `[part="value-link"][aria-pressed="true"] [part="value-label"]:contains("${text}")`
+      `[part~="value-link"][part~="value-link-selected"][aria-pressed="true"] [part="value-label"]:contains("${text}")`
     );
   },
   idleLinkValueLabel() {
     return this.shadow().find(
-      '[part="value-link"][aria-pressed="false"] [part="value-label"]'
+      '[part~="value-link"]:not([part~="value-link-selected"])[aria-pressed="false"] [part="value-label"]'
     );
   },
   selectedBoxValue() {
-    return this.shadow().find('[part="value-box"][aria-pressed="true"]');
+    return this.shadow().find(
+      '[part~="value-box"][part~="value-box-selected"][aria-pressed="true"]'
+    );
   },
   idleBoxValue() {
-    return this.shadow().find('[part="value-box"][aria-pressed="false"]');
+    return this.shadow().find(
+      '[part~="value-box"]:not([part~="value-box-selected"])[aria-pressed="false"]'
+    );
   },
   selectedBoxValueWithText(text: string) {
     return this.shadow().find(
-      `[part="value-box"][aria-pressed="true"]:contains("${text}")`
+      `[part~="value-box"][part~="value-box-selected"][aria-pressed="true"]:contains("${text}")`
     );
   },
   idleBoxValueLabel() {