+++ packages/headless/src/state/insight-app-state.ts
@@ -19,28 +19,33 @@ import {
   CategoryFacetSection,
   SortSection,
   StaticFilterSection,
+  DidYouMeanSection,
+  TabSection,
 } from './state-sections';
 
-export type InsightSearchParametersState = QuerySection &
-  FacetSection &
+export type InsightSearchParametersState = FacetSection &
   DateFacetSection &
   NumericFacetSection &
-  PaginationSection;
+  CategoryFacetSection &
+  QuerySection &
+  TabSection &
+  SortSection;
 
 export type InsightAppState = InsightSearchParametersState &
+  SearchHubSection &
   ConfigurationSection &
   VersionSection &
   InsightConfigurationSection &
   SearchSection &
-  SearchHubSection &
+  PaginationSection &
   InsightInterfaceSection &
   InsightCaseContextSection &
   ResultPreviewSection &
   FacetSection &
   FacetSearchSection &
+  FacetOptionsSection &
   QuerySetSection &
   QuerySuggestionSection &
-  CategoryFacetSection &
-  FacetOptionsSection &
   SortSection &
-  StaticFilterSection;
+  StaticFilterSection &
+  DidYouMeanSection;